+++ frontend/src/Navigation.tsx
@@ -1,7 +1,7 @@
 /*
   This file is part of Edgehog.
 
-  Copyright 2021-2022 SECO Mind Srl
+  Copyright 2021-2023 SECO Mind Srl
 
   Licensed under the Apache License, Version 2.0 (the "License");
   you may not use this file except in compliance with the License.
@@ -36,6 +36,9 @@ enum Route {
   hardwareTypes = "/hardware-types",
   hardwareTypesNew = "/hardware-types/new",
   hardwareTypesEdit = "/hardware-types/:hardwareTypeId/edit",
+  baseImageCollections = "/base-image-collections",
+  baseImageCollectionsNew = "/base-image-collections/new",
+  baseImageCollectionsEdit = "/base-image-collections/:baseImageCollectionId/edit",
   login = "/login",
   logout = "/logout",
 }
@@ -57,6 +60,12 @@ type ParametricRoute =
   | { route: Route.hardwareTypes }
   | { route: Route.hardwareTypesNew }
   | { route: Route.hardwareTypesEdit; params: { hardwareTypeId: string } }
+  | { route: Route.baseImageCollections }
+  | { route: Route.baseImageCollectionsNew }
+  | {
+      route: Route.baseImageCollectionsEdit;
+      params: { baseImageCollectionId: string };
+    }
   | { route: Route.login }
   | { route: Route.logout };